+++ src/views/List/Backup/SetBackupWarning.tsx
@@ -1,24 +1,18 @@
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { setBackupCheckedItems } from '../../../features/checkedItems/checkedItemsSlice';
 import { setBackupGroups } from '../../../features/groups/groupsSlice';
 import { setBackupStockList } from '../../../features/stockList/stockListSlice';
-import {
-  closeStockModal,
-  initialStockModal,
-  selectModalProps,
-} from '../../../features/stockModal/stockModalSlice';
+import { initialStockModal } from '../../../features/stockModal/stockModalSlice';
 import WarningModal from '../../../components/WarningModal';
 
-export type SetBackupWarningProps = {
+type Props = {
+  onClose: () => void;
   data: any;
 };
 
-const SetBackupWarning = () => {
+const SetBackupWarning = ({ onClose, data }: Props) => {
   const dispatch = useDispatch();
-  const { data } = useSelector(
-    selectModalProps('SetBackupWarning'),
-  ) as SetBackupWarningProps;
 
   const backupData = () => {
     if (data === null || data === undefined) return;
@@ -27,8 +21,6 @@ const SetBackupWarning = () => {
     dispatch(setBackupStockList(data.stockList));
     dispatch(initialStockModal());
   };
-
-  const onClose = () => dispatch(closeStockModal('SetBackupWarning'));
 
   return (
     <WarningModal
